+++ src/adapters/express.ts
@@ -1,0 +1,219 @@
+import express from 'express';
+import path from 'path';
+
+import * as H5P from '../';
+
+export default function(
+    h5pEditor: H5P.H5PEditor,
+    h5pCorePath: string
+): express.Router {
+    const router = express.Router();
+
+    router.get(`/libraries/:uberName/:file(*)`, async (req, res) => {
+        const stream = h5pEditor.libraryManager.getFileStream(
+            H5P.LibraryName.fromUberName(req.params.uberName),
+            req.params.file
+        );
+        stream.on('end', () => {
+            res.end();
+        });
+        stream.pipe(res.type(path.basename(req.params.file)));
+    });
+
+    router.get(`/content/:id/:file(*)`, async (req, res) => {
+        const stream = await h5pEditor.getContentFileStream(
+            req.params.id,
+            req.params.file,
+            req.user
+        );
+        stream.on('end', () => {
+            res.end();
+        });
+        stream.pipe(res.type(path.basename(req.params.file)));
+    });
+
+    router.get(
+        `${h5pEditor.config.temporaryFilesPath}/:file(*)`,
+        async (req, res) => {
+            const stream = await h5pEditor.getContentFileStream(
+                undefined,
+                req.params.file,
+                req.user
+            );
+            stream.on('end', () => {
+                res.end();
+            });
+            stream.pipe(res.type(path.basename(req.params.file)));
+        }
+    );
+
+    router.use('/', express.static(h5pCorePath));
+
+    router.get('/:contentId/render', (req, res) => {
+        if (!req.params.contentId) {
+            return res.redirect('/');
+        }
+
+        const libraryLoader = (lib, maj, min) =>
+            h5pEditor.libraryManager.loadLibrary(
+                new H5P.LibraryName(lib, maj, min)
+            );
+        Promise.all([
+            h5pEditor.contentManager.loadContent(req.query.contentId, req.user),
+            h5pEditor.contentManager.loadH5PJson(req.query.contentId, req.user)
+        ]).then(([contentObject, h5pObject]) =>
+            new H5P.H5PPlayer(libraryLoader as any, {}, null, null, null)
+                .render(req.query.contentId, contentObject, h5pObject)
+                .then(h5pPage => res.end(h5pPage))
+                .catch(error => res.status(500).end(error.message))
+        );
+    });
+
+    router.get('/download', async (req, res) => {
+        if (!req.query.contentId) {
+            return res.redirect('/');
+        }
+
+        const packageExporter = new H5P.PackageExporter(
+            h5pEditor.libraryManager,
+            h5pEditor.translationService,
+            h5pEditor.config,
+            h5pEditor.contentManager
+        );
+
+        // set filename for the package with .h5p extension
+        res.setHeader(
+            'Content-disposition',
+            `attachment; filename=${req.query.contentId}.h5p`
+        );
+        await packageExporter.createPackage(req.query.contentId, res, req.user);
+    });
+
+    router.get('/params', (req, res) => {
+        h5pEditor
+            .loadH5P(req.query.contentId)
+            .then(content => {
+                res.status(200).json(content);
+            })
+            .catch(() => {
+                res.status(404).end();
+            });
+    });
+
+    router.get('/edit', async (req, res) => {
+        h5pEditor.render(req.query.contentId).then(page => res.end(page));
+    });
+
+    router.post('/edit', (req, res) => {
+        h5pEditor
+            .saveH5P(
+                req.query.contentId,
+                req.body.params.params,
+                req.body.params.metadata,
+                req.body.library,
+                req.ser
+            )
+            .then(() => {
+                res.status(200).end();
+            });
+    });
+
+    router.get('/ajax', (req, res) => {
+        const { action } = req.query;
+        const { majorVersion, minorVersion, machineName, language } = req.query;
+
+        switch (action) {
+            case 'content-type-cache':
+                h5pEditor
+                    .getContentTypeCache(req.user)
+                    .then(contentTypeCache => {
+                        res.status(200).json(contentTypeCache);
+                    });
+                break;
+
+            case 'libraries':
+                h5pEditor
+                    .getLibraryData(
+                        machineName,
+                        majorVersion,
+                        minorVersion,
+                        language
+                    )
+                    .then(library => {
+                        res.status(200).json(library);
+                    });
+                break;
+
+            default:
+                res.status(400).end();
+                break;
+        }
+    });
+
+    router.post('/ajax', async (req, res) => {
+        const { action } = req.query;
+        switch (action) {
+            case 'libraries':
+                const libraryOverview = await h5pEditor.getLibraryOverview(
+                    req.body.libraries
+                );
+                res.status(200).json(libraryOverview);
+                break;
+            case 'translations':
+                const translationsResponse = await h5pEditor.getLibraryLanguageFiles(
+                    req.body.libraries,
+                    req.query.language
+                );
+                res.status(200).json({
+                    data: translationsResponse,
+                    success: true
+                });
+                break;
+            case 'files':
+                const uploadFileResponse = await h5pEditor.saveContentFile(
+                    req.body.contentId === '0'
+                        ? req.query.contentId
+                        : req.body.contentId,
+                    JSON.parse(req.body.field),
+                    req.files.file,
+                    req.user
+                );
+                res.status(200).json(uploadFileResponse);
+                break;
+            case 'library-install':
+                await h5pEditor.installLibrary(req.query.id, req.user);
+                const contentTypeCache = await h5pEditor.getContentTypeCache(
+                    req.user
+                );
+                res.status(200).json({
+                    data: contentTypeCache,
+                    success: true
+                });
+                break;
+            case 'library-upload':
+                const contentId = await h5pEditor.uploadPackage(
+                    req.files.h5p.data,
+                    req.query.contentId,
+                    req.user
+                );
+                const [content, contentTypes] = await Promise.all([
+                    h5pEditor.loadH5P(contentId),
+                    h5pEditor.getContentTypeCache(req.user)
+                ]);
+                res.status(200).json({
+                    data: {
+                        content: content.params.params,
+                        contentTypes,
+                        h5p: content.h5p
+                    },
+                    success: true
+                });
+                break;
+            default:
+                res.status(500).end('NOT IMPLEMENTED');
+                break;
+        }
+    });
+
+    return router;
+}
